Stricter test for names of interpolated variables
The old code allowed any mix of [\w:]. This tripped over "$foo::$bar",
in which the correct parse has interpolation '$foo', string '::', and
interpolation '$bar', but the former code made this interpolations
'$foo::' and '$bar'. The correct parse is verified by B::Deparse, which
deparses "$foo::$bar" to "${foo}::$bar".