Eliminte blockers to running under Perl 5.6.
The significant change was replacing (?-1) in a regular expression
(introduced in 5.9.5) with (??{...}) (going back to Heaven knows when,
and used in Regexp::Common back in 2003. The rest of the changes were
correcting stray 'use 5.008;' statements. I also added
xt/author/minimum_perl.t (based on Perl::MinimumVersion) to try to make
sure nothing untoward sneaks back in.