Only standalone grapheme delims starting w/ 5.29.0
This actually means significantly expanding the delimiters allowed, and
then having perl_version_removed() return '5.029' if the delimiter
matches /\p{Mark}/. This code does not operate under Perl 5.6 because it
can not compile the regular expression. Whether PPI will actually
generate a token with such delimiters is unknown to me.

Collateral with this, accept word characters as delimiters, but only
with at least one space between the operator and the expression -- that
is, 'qq xyx' is OK, but 'qqxyx' is not.