Normalize interpolation before feeding to PPI.
This produces a more-easily-analyzed PPI parse.

The code was actually put in PPIx::QuoteLike::Utils with the intent of
sharing it with PPIx::Regexp. It dragged a bunch of stuff with it from
PPIx::QuoteLike.

The ->variables() method is now discouraged. It was written to support
Perl::Critic::Policy::Variables::ProhibitUnusedVarsStricter, but turned
out to be inadequate for the job.